Rename redis.js to redisclient.js
