+++ redisclient.js
